2021-04-21.04:42pm

+++ src/eco/controller/NoticeDeleteServlet.java
@@ -3,6 +3,7 @@ package eco.controller;
 import java.io.IOException;
 import java.io.PrintWriter;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.Servlet;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,7 +11,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import member.model.service.MemberService;
+import member.model.vo.Member;
 import notice.model.service.NoticeService;
+import notice.model.vo.Notice;
 
 @WebServlet("/eco/notice/delete")
 public class NoticeDeleteServlet extends HttpServlet implements Servlet {
@@ -23,13 +27,49 @@ public class NoticeDeleteServlet extends HttpServlet implements Servlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		int noticeNo = Integer.parseInt(request.getParameter("noticeNo"));
+		
+		String memberId = (String)request.getSession().getAttribute("memberId");
+		
+		if (memberId == null) {
+			response.setContentType("text/html; charset=utf-8");
+			PrintWriter out = response.getWriter();
+			String msg = "로그인을 해주세요."; // 오류 메세지
+			out.println("<script>");
+			out.println("alert('" + msg + "');");
+			out.println("history.back();");
+			out.println("</script>");
+			out.flush();
+			out.close();
+			return;
+		}
+		
+		// 공지사항 번호로 공지사항을 가져옴
+		Notice notice = new NoticeService().printOnebyNo(noticeNo);
+		System.out.println(notice.getMbId());
+		
+		// 공지사항 작성자가 로그인한 사용자와 일치하는지 확인
+		if (notice.getMbId() != memberId) {
+			response.setContentType("text/html; charset=utf-8");
+			PrintWriter out = response.getWriter();
+			String msg = "삭제할 권한이 없습니다."; // 오류 메세지
+			out.println("<script>");
+			out.println("alert('" + msg + "');");
+			out.println("history.back();");
+			out.println("</script>");
+			out.flush();
+			out.close();
+			return;
+		}
+		
 		int result = new NoticeService().deleteNotice(noticeNo);
-		if (result > 0) {
-			response.sendRedirect("/eco/notice/list");
+		
+		if (0 < result) {
+			RequestDispatcher view = request.getRequestDispatcher("/WEB-INF/views/eco/noticeList.jsp");
+			view.forward(request, response);
 		} else {
 			response.setContentType("text/html; charset=utf-8");
 			PrintWriter out = response.getWriter();
-			String msg = "표시할 내용이 없습니다."; // 오류 메세지
+			String msg = "삭제에 실패하였습니다."; // 오류 메세지
 			out.println("<script>");
 			out.println("alert('" + msg + "');");
 			out.println("history.back();");
